Rename file inconsistencies
