+++ packages/ludiek/src/engine/LudiekEngineConfig.ts
@@ -1,11 +1,11 @@
 import { LudiekEvaluator } from '@ludiek/engine/condition/LudiekEvaluator';
 import { LudiekConsumer } from '@ludiek/engine/input/LudiekConsumer';
 import { LudiekProducer } from '@ludiek/engine/output/LudiekProducer';
-import { LudiekController } from './request/LudiekController';
+import { LudiekController } from '@ludiek/engine/request/LudiekController';
 import { LudiekPlugin } from '@ludiek/engine/LudiekPlugin';
-import { LudiekModifier } from '@ludiek/engine/modifier/LudiekModifier';
+import { LudiekModifier } from '@ludiek/engine/bonus/LudiekModifier';
 import { LudiekFeature } from '@ludiek/engine/LudiekFeature';
-import { LudiekContent } from './LudiekContent';
+import { LudiekContent } from '@ludiek/engine/LudiekContent';
 
 export interface LudiekEngineConfig<
   Plugins extends readonly LudiekPlugin[] = [],
