+++ packages/ludiek/src/engine/LudiekEngineContribution.ts
@@ -4,10 +4,10 @@ import { EngineNotInjectedError } from '@ludiek/engine/LudiekError';
 import { LudiekEvaluator } from '@ludiek/engine/condition/LudiekEvaluator';
 import { LudiekConsumer } from '@ludiek/engine/input/LudiekConsumer';
 import { LudiekProducer } from '@ludiek/engine/output/LudiekProducer';
-import { LudiekController } from './request/LudiekController';
-import { LudiekModifier } from '@ludiek/engine/modifier/LudiekModifier';
+import { LudiekController } from '@ludiek/engine/request/LudiekController';
+import { LudiekModifier } from '@ludiek/engine/bonus/LudiekModifier';
 import { LudiekFeature } from '@ludiek/engine/LudiekFeature';
-import { LudiekContent } from './LudiekContent';
+import { LudiekContent } from '@ludiek/engine/LudiekContent';
 
 export interface LudiekDependencies {
   plugins?: readonly LudiekPlugin[];
