+++ packages/ludiek/src/engine/LudiekFeature.ts
@@ -1,5 +1,5 @@
 import { LudiekElement } from '@ludiek/engine/LudiekElement';
-import { LudiekDependencies } from './LudiekEngineContribution';
+import { LudiekDependencies } from '@ludiek/engine/LudiekEngineContribution';
 
 export abstract class LudiekFeature<
   Dependencies extends LudiekDependencies = object,
