+++ packages/ludiek/src/index.ts
@@ -26,7 +26,7 @@ export { LudiekProducer, type BaseOutput, type LudiekOutput } from '@ludiek/engi
 export type { LudiekTransaction } from '@ludiek/engine/transaction/LudiekTransaction';
 
 // Engine.Requests
-export { LudiekController, type BaseRequest, type LudiekRequest } from './engine/request/LudiekController';
+export { LudiekController, type BaseRequest, type LudiekRequest } from '@ludiek/engine/request/LudiekController';
 
 // Engine.Modifiers
 export {
@@ -34,10 +34,10 @@ export {
   type BaseBonus,
   type LudiekBonus,
   type LudiekBonusContribution,
-} from '@ludiek/engine/modifier/LudiekModifier';
+} from '@ludiek/engine/bonus/LudiekModifier';
 
 // Engine.Types
-export type { LudiekDependencies, DependencyEngine } from './engine/LudiekEngineContribution';
+export type { LudiekDependencies, DependencyEngine } from '@ludiek/engine/LudiekEngineContribution';
 
 // Engine.Errors
 export { LudiekError } from '@ludiek/engine/LudiekError';
